Testes edição e página de edição com sucesso concluídos
Closes #99
